src/requirements.txt
requests
Pillow
colorama